Tutorial

EvoDynamics.jl's basic usage

First, define your model parameters. Below is a set of random parameters. See API for a description of each parameter.

using Random
import LinearAlgebra: Symmetric

nphenotypes = (4, 5)
ngenes = (7, 8)
ploidy = (2, 1)
# choose random values for epistasis matrix epistasisMat, but make sure the diagonal is
# 1.0, meaning that each locus affects itself 100%.
epistasisMat = Tuple([Random.rand(-0.5:0.01:0.5, i, i) for i in (ngenes .* ploidy)])
for index in 1:length(epistasisMat)
  for diag in 1:size(epistasisMat[index], 1)
    epistasisMat[index][diag, diag] = 1
  end
end

parameters = Dict(
  :ngenes => ngenes .* ploidy,
  :nphenotypes => nphenotypes,
  :growthrates => (0.8, 0.12),
  :interactionCoeffs => rand(-0.1:0.01:0.1, 2, 2),
  :pleiotropyMat => (rand([true, false], nphenotypes[1], ngenes[1] * ploidy[1]), rand([true, false], nphenotypes[2], ngenes[2] * ploidy[2])),
  :epistasisMat =>  epistasisMat,
  :expressionArrays => Tuple([rand() for el in 1:l] for l in ngenes .* ploidy),
  :selectionCoeffs => (0.5, 0.5),
  :ploidy => ploidy,
  :optPhenotypes => Tuple([randn(Float16, n) for n in nphenotypes]),
  :covMat => Tuple([Symmetric(rand(Float16, i[1], i[2])) for i in zip(nphenotypes, nphenotypes)]),
  :mutProbs => Tuple([(0.02, 0.0, 0.0), (0.02, 0.0, 0.0)]),
  :mutMagnitudes => Tuple([(0.05, 0.0, 0.01), (0.05, 0.0, 0.01)]),
  :N => Dict(1 => (1000, 1000)),
  :K => Dict(1 => [1000, 1000], 2 => [1000, 1000], 3 => [1000, 1000], 4 => [1000, 1000]),
  :migration_rates => [[1.0 0.02 0.02 0.02; 0.03 1.0 0.03 0.03; 0.01 0.01 1.0 0.01; 0.01 0.01 0.01 1.0] for i in 1:2],
  :E => (0.01, 0.01),
  :generations => 5,
  :space => (2,2),
  :moore => false
)

Dict{Symbol,Any} with 20 entries:
  :ploidy            => (2, 1)
  :expressionArrays  => ([0.137556, 0.327311, 0.926159, 0.282935, 0.506852, 0.0…
  :interactionCoeffs => [-0.03 0.1; 0.04 0.02]
  :migration_rates   => [[1.0 0.02 0.02 0.02; 0.03 1.0 0.03 0.03; 0.01 0.01 1.0…
  :space             => (2, 2)
  :moore             => false
  :epistasisMat      => ([1.0 -0.32 … -0.36 -0.43; 0.22 1.0 … -0.03 -0.31; … ; …
  :selectionCoeffs   => (0.5, 0.5)
  :K                 => Dict(4=>[1000, 1000],2=>[1000, 1000],3=>[1000, 1000],1=…
  :generations       => 5
  :E                 => (0.01, 0.01)
  :growthrates       => (0.8, 0.12)
  :N                 => Dict(1=>(1000, 1000))
  :mutMagnitudes     => ((0.05, 0.0, 0.01), (0.05, 0.0, 0.01))
  :covMat            => (Float16[0.2324 0.742 0.10156 0.4883; 0.742 0.3545 0.44…
  :ngenes            => (14, 8)
  :pleiotropyMat     => (Bool[1 1 … 0 0; 1 0 … 0 0; 1 0 … 1 0; 0 0 … 1 1], Bool…
  :mutProbs          => ((0.02, 0.0, 0.0), (0.02, 0.0, 0.0))
  :optPhenotypes     => (Float16[-1.895, -0.962, -0.1774, -0.3193], Float16[1.2…
  :nphenotypes       => (4, 5)

We can the use the runmodel function to create a model from these parameters and run the simulation.

EvoDynamics.runmodel — Function

runmodel(parameters::Dict; kwargs)

Creates and runs a model given parameters. Returns a DataFrame of collected data, which are specified by kwargs.

Keywords

adata=[] agent data to be collected. Either agent fields or functions that accept an agent as input can be put in the array. To aggregate collected data, provide tuples inside the array. For example, to collect mean and median fitness of individuals which is in field W, your array will be [(:W,mean), (:W,median)].

mdata=[meanfitnessper_species] model data to be collected. By default, collects mean population fitness per species. Each row of the output DataFrame corresponds to all agents and each column is the value function applied to a field. The functions in the array are applied to the model object.

when::AbstractArray{Int}=1:parameters[:generations] The generations from which data are collected

replicates::Int = 0 Number of replicates per simulation.

parallel::Bool = false Whether to run replicates in parallel. If true, you should add processors to your julia session (e.g. by addprocs(n)) and define your parameters and EvoDynamics on all workers. To do that, add @everywhere before them. For example, @everywhere EvoDynamics.

using EvoDynamics
data, model = runmodel(parameters)
data[1:5, :]