Tutorial

EvoDynamics.jl's basic usage

First, define your model parameters in a YAML file (here, we call it parameters.yml). Simple Wright-Fisher and Predator prey have examples of initiation parameters. See API for a description of each parameter.

We can the use the runmodel function to create a model from these parameters and run the simulation.

EvoDynamics.runmodel — Function

runmodel(param_file::AbstractString; kwargs)

Creates and runs a model given parameters. Returns a DataFrame of collected data, which are specified by kwargs.

Keywords

adata=[] agent data to be collected. Either agent fields or functions that accept an agent as input can be put in the array. To aggregate collected data, provide tuples inside the array. For example, to collect mean and median fitness of individuals which is in field W, your array will be [(:W,mean), (:W,median)].

mdata=[meanfitnessper_species] model data to be collected. By default, collects mean population fitness per species. Each row of the output DataFrame corresponds to all agents and each column is the value function applied to a field. The functions in the array are applied to the model object.

when=nothing The generations from which data are collected. By default collect at all steps.

replicates::Int = 0 Number of replicates per simulation.

parallel::Bool = false Whether to run replicates in parallel. If true, you should add processors to your julia session (e.g. by addprocs(n)) and define your parameters and EvoDynamics on all workers. To do that, add @everywhere before them. For example, @everywhere EvoDynamics. TODO: use ensemblerun! instead of run.

using EvoDynamics
agentdata, modeldata, model = runmodel("parameters.yml")

For data collection, see API.

Creating simulation parameter files

EvoDynamics.jl reads simulation parameters (API) from a human-readable YAML file. This file can be populated manually using a text editor or from within a Julia session.

To create parameters from within a Julia session and write them to a YAML file, you can follow the example below.

To define species parameters, create a dictionary whose keys are numbers starting from 1 for the number of species you have, and values are dictionaries themselves with the species parameter names and values. In the example below, for brevity, I only add a few parameters for each species (see API for the complete list of parameters).

species_params = Dict(
  1=>Dict("name" => "a", "number of genes" => 2, "number of phenotypes" => 2, "abiotic phenotypes" => [1]),
  2=>Dict("name" => "b", "number of genes" => 2, "number of phenotypes" => 2, "abiotic phenotypes" => [1])
)

To define model parameters, create a new dictionary whose keys are the parameter names and values the parameter values. Here, again I only define a few parameters for brevity.

model_params = Dict("generations"=> 100, "space"=> [6,10], "food sources" => [1.0, 0.7, 0.0, 0.0])

Finally, create a dictionary mixing the two dictionaries before with keys "species" and "model".

data = Dict("species"=> species_params, "model"=> model_params)

This dictionary can be written as a YAML file in the correct format.

using YAML
f = "params.yml"
YAML.write_file(f, data)