Tutorial

EvoDynamics.jl's basic usage

First, define your model parameters. Below is a set of random parameters. See API for a description of each parameter.

using Random
import LinearAlgebra: Symmetric

P = (4, 5)
L = (7, 8)
m = (2, 1)
# choose random values for epistasis matrix A, but make sure the diagonal is
# 1.0, meaning that each locus affects itself 100%.
A = Tuple([Random.rand(-0.5:0.01:0.5, i, i) for i in (L .* m)])
for index in 1:length(A)
  for diag in 1:size(A[index], 1)
    A[index][diag, diag] = 1
  end
end

parameters = Dict(
  :L => L .* m,
  :P => P,
  :R => (0.8, 0.12),
  :C => rand(-0.1:0.01:0.1, 2, 2),
  :B => (rand([true, false], P[1], L[1] * m[1]), rand([true, false], P[2], L[2] * m[2])),
  :A =>  A,
  :Q => Tuple([rand() for el in 1:l] for l in L .* m),
  :Y => (0.5, 0.5),
  :m => m,
  :T => Tuple([randn(Float16, n) for n in P]),
  :Ω => Tuple([Symmetric(rand(Float16, i[1], i[2])) for i in zip(P, P)]),
  :M => Tuple([(0.02, 0.0, 0.0), (0.02, 0.0, 0.0)]),
  :D => Tuple([(0.05, 0.0, 0.01), (0.05, 0.0, 0.01)]),
  :N => Dict(1 => (1000, 1000)),
  :K => Dict(1 => [1000, 1000], 2 => [1000, 1000], 3 => [1000, 1000], 4 => [1000, 1000]),
  :migration_rates => [[1.0 0.02 0.02 0.02; 0.03 1.0 0.03 0.03; 0.01 0.01 1.0 0.01; 0.01 0.01 0.01 1.0] for i in 1:2],
  :E => (0.01, 0.01),
  :generations => 5,
  :space => (2,2),
  :moore => false
)

Dict{Symbol,Any} with 20 entries:
  :D               => ((0.05, 0.0, 0.01), (0.05, 0.0, 0.01))
  :migration_rates => Array{Float64,2}[[1.0 0.02 0.02 0.02; 0.03 1.0 0.03 0.03;…
  :B               => (Bool[0 1 … 1 1; 1 0 … 1 1; 0 0 … 1 1; 1 1 … 1 1], Bool[0…
  :space           => (2, 2)
  :moore           => false
  :Ω               => (Float16[0.125 0.536 0.02832 0.5; 0.536 0.716 0.01563 0.6…
  :K               => Dict(4=>[1000, 1000],2=>[1000, 1000],3=>[1000, 1000],1=>[…
  :generations     => 5
  :E               => (0.01, 0.01)
  :R               => (0.8, 0.12)
  :M               => ((0.02, 0.0, 0.0), (0.02, 0.0, 0.0))
  :A               => ([1.0 -0.13 … -0.01 -0.04; -0.05 1.0 … 0.03 -0.13; … ; 0.…
  :N               => Dict(1=>(1000, 1000))
  :Q               => ([0.0846451, 0.756579, 0.718204, 0.277072, 0.308461, 0.28…
  :Y               => (0.5, 0.5)
  :T               => (Float16[-1.965, 0.9927, -0.0001448, 2.11], Float16[0.465…
  :P               => (4, 5)
  :m               => (2, 1)
  :L               => (14, 8)
  :C               => [0.06 0.09; -0.01 0.02]

We can the use the runmodel function to create a model from these parameters and run the simulation.

EvoDynamics.runmodel — Function

runmodel(parameters::Dict; kwargs)

Creates and runs a model given parameters. Returns a DataFrame of collected data, which are specified by kwargs.

Keywords

collect::Dict=Dict(:model => [meanfitnessper_species]) Data to be collected. By default, collects mean population fitness per species.

when::AbstractArray{Int}=1:parameters[:generations] The generations from which data are collected

replicates::Int = 0 Number of replicates per simulation.

parallel::Bool = false Whether to run replicates in parallel. If true, you should add processors to your julia session (e.g. by addprocs(n)) and define your parameters and EvoDynamics on all workers. To do that, add @everywhere before them. For example, @everywhere EvoDynamics.

using EvoDynamics
data = runmodel(parameters)
data[1:5, :]

step	mean_fitness_per_species(model)_1	mean_fitness_per_species(model)_2

Int64	Float32	Float32

5 rows × 3 columns

1	0	100000.0	4.10193e-17

2	1	100000.0	1.76985e-16

3	2	100000.0	9.51349e-16

4	3	100000.0	8.73462e-13

5	4	100000.0	2.16793e-12