Tutorial

EvoDynamics.jl's basic usage

First, define your model parameters. Here is a set of random parameters:

using Random
import LinearAlgebra: Symmetric

P = (4, 5)  # a tuple  specifying the number of traits p for each species
L = (7, 8)  # a tuple  specifying the number of loci l for each species
m = 1  # ploidy, m=2 diploid. Currently only haploids are implemented
parameters = Dict(
  :L => L,
  :P => P,
  :B =>  Tuple([Random.bitrand(i[1], i[2]) for i in zip(P, L)]),  # a tuple  of pleiotropy matrices, one for each species. Each matrix consists of zeros and ones only. make sure no rows are all zero (a trait is not controled :by any locus)
  :γ => (-0.5, -0.5), # a tuple  of selection coefficients for each species
  :m => m,
  :T => Tuple([randn(Float16, n) for n in P]), # a tuple  of arrays, each inner array θ specifying optimal phenotypes for each species
  :Ω => Tuple([Symmetric(rand(Float16, i[1], i[2])) for i in zip(P, P)]), # a tuple  of matrices, each of wich ω represents a covariance matrix of the selection surface
  :M => (0.02, 0.02), # a tuple of mutation rates μ for each species
  :MB => (0.05, 0.05), # a tuple of mutation rates μ<sub>B</sub> for each species
  :N => (1000, 1000), # a tuple  for population size of each species
  :Y => Tuple([rand(Float16, i*m) for i in L]), # a tuple  of Arrays, each specifying the initial y vector of each species
  :E => (0.8, 0.8), # a tuple  of the variance of a normal distribution ε representing environmental noise for each species.
  :generations => 100 # number of generations to run the simulation
)

Dict{Symbol,Any} with 13 entries:
  :B           => (Bool[0 0 … 1 1; 0 0 … 1 1; 1 0 … 0 1; 1 1 … 0 0], Bool[0 0 ……
  :MB          => (0.05, 0.05)
  :Ω           => (Float16[0.05957 0.4736 0.9814 0.2822; 0.4736 0.37 0.8604 0.0…
  :generations => 100
  :E           => (0.8, 0.8)
  :N           => (1000, 1000)
  :M           => (0.02, 0.02)
  :Y           => (Float16[0.6533, 0.4043, 0.9307, 0.6914, 0.5527, 0.589, 0.059…
  :γ           => (-0.5, -0.5)
  :P           => (4, 5)
  :m           => 1
  :T           => (Float16[-0.3242, -0.6206, 0.944, -0.2974], Float16[0.06, 0.8…
  :L           => (7, 8)

We can the use the runmodel function to create a model from these parameters and run the simulation.

EvoDynamics.runmodel — Function

runmodel(parameters::Dict; kwargs)

Creates and runs a model given parameters. Returns a DataFrame of collected data, which are specified by kwargs.

Keywords

collect::Dict=Dict(:model => [meanfitnessper_species]) Data to be collected. By default, collects mean population fitness per species.

when::AbstractArray{Int}=1:parameters[:generations] The generations from which data are collected

replicates::Int = 0 Number of replicates per simulation.

parallel::Bool = false Whether to run replicates in parallel. If true, you should add processors to your julia session (e.g. by addprocs(n)) and define your parameters and EvoDynamics on all workers. To do that, add @everywhere before them. For example, @everywhere EvoDynamics.

using EvoDynamics
data = runmodel(parameters)
data[1:5, :]

step	mean_fitness_per_species(model)_1	mean_fitness_per_species(model)_2

Int64	Float32	Float32

5 rows × 3 columns

1	0	5.61052	1.00089

2	1	0.0552727	0.000451935

3	2	0.0498567	2.20496e-7

4	3	0.0626355	NaN

5	4	0.0856263	NaN