Tutorial

Installation

Install using the following command in a Julia REPL.

]add EvoDynamics

Basic usage

Parameters of a model should be put in a YAML file with the structure below. Note that spaces and indentations are meaningful in YAML. Indentations should be spaces not tabs.

See Simple Wright-Fisher and Predator prey for complete examples of parameter files.

species:
  1:
    name: a
    parameter 1: ...
    parameter 2: ...
    ...
  2:
    name: b
    parameter 1: ... 
    parameter 2: ... 
    ...
model:
  model parameter 1: ...
  model parameter 2: ...
  ...

The file has two main levels: species and model. species stores species specific parameters as many different species as you want.

Since we cannot write a matrix in a YAML file, any parameter that is a matrix should be converted to a vector. In Julia, you can do this by vec(yourmatrix).

First, define your model parameters in a YAML file (here, we call it parameters.yml). Simple Wright-Fisher and Predator prey have examples of initiation parameters. See Model description for a description of each parameter.

We can the use the runmodel function to create a model from these parameters and run the simulation.

EvoDynamics.runmodel — Function

runmodel(param_file::AbstractString; kwargs)

Creates and runs a model given parameters. Returns a DataFrame of collected data, which are specified by kwargs.

Keywords

adata=[] agent data to be collected. Either agent fields or functions that accept an agent as input can be put in the array. To aggregate collected data, provide tuples inside the array. For example, to collect mean and median fitness of individuals which is in field W, your array will be [(:W,mean), (:W,median)].

mdata=[meanfitnessper_species] model data to be collected. By default, collects mean population fitness per species. Each row of the output DataFrame corresponds to all agents and each column is the value function applied to a field. The functions in the array are applied to the model object.

when=nothing The generations from which data are collected. By default collect at all steps.

replicates::Int = 0 Number of replicates per simulation.

parallel::Bool = false Whether to run replicates in parallel. If true, you should add processors to your julia session (e.g. by addprocs(n)) and define your parameters and EvoDynamics on all workers. To do that, add @everywhere before them. For example, @everywhere EvoDynamics.

seeds = optionally, provide an array of integers as seeds for each replicate.

using EvoDynamics
agentdata, modeldata, model = runmodel("parameters.yml")

Creating simulation parameter files

EvoDynamics.jl reads simulation parameters (Model description) from a human-readable YAML file. This file can be populated manually using a text editor or from within a Julia session.

To create parameters from within a Julia session and write them to a YAML file, you can follow the example below.

To define species parameters, create a dictionary whose keys are numbers starting from 1 for the number of species you have, and values are dictionaries themselves with the species parameter names and values. In the example below, for brevity, I only add a few parameters for each species (see Model description for the complete list of parameters).

species_params = Dict(
  1=>Dict("name" => "a", "number of genes" => 2, "number of phenotypes" => 2, "abiotic phenotypes" => [1]),
  2=>Dict("name" => "b", "number of genes" => 2, "number of phenotypes" => 2, "abiotic phenotypes" => [1])
)

To define model parameters, create a new dictionary whose keys are the parameter names and values the parameter values. Here, again I only define a few parameters for brevity.

model_params = Dict("generations"=> 100, "space"=> [6,10], "food sources" => [1.0, 0.7, 0.0, 0.0])

Finally, create a dictionary mixing the two dictionaries before with keys "species" and "model".

data = Dict("species"=> species_params, "model"=> model_params)

This dictionary can be written as a YAML file in the correct format.

using YAML
f = "params.yml"
YAML.write_file(f, data)

Collecting data

The interface to the model is from the runmodel function (see Tutorial).

EvoDynamics.jl uses Agents.jl underneath. See Agents.jl's documentation for details about writing functions to collect any data during simulations. Here, we explain the specific implementation of the model.

There are two main objects from which you can collect data: and agent object of type AbstractAgent and a model object of type ABM. Both of these types are defined the Agents.jl package.

Agent object has the following fields that are individual specific: id, pos, species, epistasisMat (epistasis matrix), pleiotropyMat (pleiotropy matrix), q (gene expression array), biotic_phenotype, abiotic_phenotype, age, sex, energy, interaction_history, and W (fitness).

The model object has the following fields that can be accessed with the . syntax and describe properties of species or the model: ngenes, nphenotypes, growthrates, selectionCoeffs, ploidy, optvals (optimal values), optinds (optval indices per generation), mutProbs (mutation probabilities), mutMagnitudes (mutation magnitudes), N, E (environmental noise), generations, nspecies, migration_traits, vision_radius, check_fraction, migration_thresholds, step, biotic_phenotypes (indices of biotic phenotypes per species), abiotic_phenotypes, max_ages, food_sources, interactions, resources, recombination, initial_energy.

You can collect data from agents and/or from the model object. To collect data from agents, use the adata keyword argument in the runmodel function, and to collect data from the model, use the mdata keyword. A complete description of the values these keywords take are at data collection section of the Agents.jl package.

For example, we use the function below to count the number of individual per species:

"Returns the size of each species."
function species_N(model::ABM)
  allagents = model.agents
  if length(allagents) == 0
    return fill(0, model.nspecies)
  else
    counts = countmap([a.species for a in values(model.agents)])
    output = fill(0, model.nspecies)
    for (k, v) in counts
      output[k] = v
    end
    return output
  end
end

using EvoDynamics

agentdata, modeldata, model = runmodel("parameters.yml", mdata=[species_N])